refactor: remove Carousel component and related styles
- Deleted the Carousel component and its associated styles from the project.
- Updated the AvailableColors component to simplify color selection logic.
- Adjusted imports in the Product component to reflect the Carousel removal.
- Added a placeholder image constant and SVG for better image handling in the ProductImageGallery.
