+++ src/features/website/Product/Carousel/styles/carousel.scss
@@ -1,4 +1,4 @@
-@import '../../../../shared/styles/variables.scss';
+@import '../../../../../shared/styles/variables.scss';
 
 .carousel {
     position: relative;
