+++ src/features/website/ProductImageGallery/ProductImageGallery.tsx
@@ -6,6 +6,7 @@ import { Box, Group, Stack } from '@mantine/core';
 import Image from 'next/image';
 
 import styles from './styles/productImageGallery.module.scss';
+import { placeHolderImageURL } from '@/shared/constants/urls';
 
 interface Props {
     images: string[];
@@ -14,15 +15,11 @@ interface Props {
 export default function ProductImageGallery({ images }: Props) {
     const [selectedImage, setSelectedImage] = useState(images[0]);
 
-    const selectImage = (img: string) => {
-        setSelectedImage(img);
-    };
-
     return (
         <Stack gap="16px">
             <Box className={styles.mainImageWrapper}>
                 <Image
-                    src={selectedImage}
+                    src={selectedImage || placeHolderImageURL}
                     alt="Selected product image"
                     fill
                     className={styles.mainImage}
@@ -36,7 +33,7 @@ export default function ProductImageGallery({ images }: Props) {
                         className={`${styles.thumbnailWrapper} ${
                             selectedImage === image ? styles.active : ''
                         }`}
-                        onClick={() => selectImage(image)}
+                        onClick={() => setSelectedImage(image)}
                     >
                         <Image
                             src={image}
